+++ flutter-glfw/src/handler.rs
@@ -1,8 +1,8 @@
+use flust_plugins::platform::{AppSwitcherDescription, MimeError, PlatformHandler};
+use flust_plugins::textinput::TextInputHandler;
+use flust_plugins::window::{PositionParams, WindowHandler};
 use flutter_engine::tasks::TaskRunnerHandler;
 use flutter_engine::FlutterOpenGLHandler;
-use flutter_plugins::platform::{AppSwitcherDescription, MimeError, PlatformHandler};
-use flutter_plugins::textinput::TextInputHandler;
-use flutter_plugins::window::{PositionParams, WindowHandler};
 use glfw::Context;
 use parking_lot::Mutex;
 use std::cell::RefCell;